requirements.txt
flask>=3.0.0
pillow>=10.0.0
werkzeug>=3.0.0
python-dotenv==1.0.1 